requirements.txt
Django==1.6.2
South==0.8.4
django-admin-tools==0.5.1
django-debug-toolbar==1.0.1
django-extensions==1.3.3
django-forms-bootstrap==3.0.0
django-datatables-view
django-simple-captcha
django-import-export
ipython==1.2.0
six==1.5.2
sqlparse==0.1.11
wsgiref==0.1.2
jdatetime